Feat: Composer library support
