+++ src/Exceptions/CurlException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Thevenrex\Curlx\Exceptions;
+
+use Exception;
+
+class CurlException extends Exception {}
