+++ src/Interfaces/CookieJarInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Thevenrex\Curlx\Interfaces;
+
+interface CookieJarInterface {
+
+    public function setFileName(string $filename);
+
+    public function getFileName(): string;
+}
